+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - `ex.TransformComponent.posChanged$` has been removed, it incurs a steep performance cost
 - `ex.EventDispatcher` meta events 'subscribe' and 'unsubscribe' were unused and undocumented and have been removed
 - `ex.TileMap` tlies are now drawn from the lower left by default to match with `ex.IsometricMap` and Tiled, but can be configured with `renderFromTopOfGraphic` to restore the previous behavior.
+- Scene `onActivate` and `onDeactivate` methods have been changed to receive a single parameter, an object containing the `previousScene`, `nextScene`, and optional `data` passed in from `goToScene()`
 
 ### Deprecated
 
@@ -114,6 +115,18 @@ This project adheres to [Semantic Versioning](http://semver.org/).
   // actor will now move to (100, 100) and rotate to Math.PI/2 at the same time!!
   ```
 - Add target element id to `ex.Screen.goFullScreen('some-element-id')` to influence the fullscreen element in the fullscreen browser API.
+- Added optional `data` parameter to `goToScene`, which gets passed to the target scene's `onActivate` method.
+  ```typescript
+  class SceneA extends ex.Scene {
+    /* ... */
+
+    onActivate(context: ex.SceneActivationContext<{ foo: string }>) {
+      console.log(context.data.foo); // bar
+    }
+  }
+
+  engine.goToScene('sceneA', { foo: 'bar' })
+  ```
 
 ### Fixed
 - Fixed bug where a deferred `goToScene` would preserve the incorrect scene so `engine.add(someActor)` would place actors in the wrong scene after transitioning to another.